Chapter04/requirements.txt
azureml-sdk==1.3.0
azureml-dataprep==1.4.3
pandas==0.23.4
numpy==1.16.2
scikit-learn==0.20.3
pyspark==2.4.3
pyarrow==0.17.0